登录账号请求逻辑

+++ Unity/Codes/Model/Demo/Account/AccountInfoComponent.cs
@@ -3,6 +3,6 @@
     public class AccountInfoComponent : Entity, IAwake, IDestroy
     {
         public string Token;
-        public int AccountId;
+        public long AccountId;
     }
 }
